Created a SpendingBreakdownCard component
